+) lib/clock_main -- CLOCK_STATE_SET_DATE
The clock now shows time and date. Both can be set. Date adjustments account
for changes in year and month for days, i.e. if it is Jan 31 2016 and the
month is increased to Feb, then days will be automatically set to 29.
Now if the year is increased to 2015, the days will be switched to 28